+++ hemle_groupwork/src/components/About/OurFriends/style.css
