+++ hemle_groupwork/src/components/About/data.jsx
@@ -1,0 +1,47 @@
+import person1 from "../../assets/home/person1.png"
+import person2 from "../../assets/home/person2.png"
+import person3 from "../../assets/home/person3.png"
+import person4 from "../../assets/home/person4.png"
+
+const data =
+{
+    titleTop: {
+        title: "Afrique football",
+        subTitle: "Since 1987"
+    },
+    whoWeAre: {
+        title: "Notre equipe de direction",
+        description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Dignissim blandit arcu nisl erat sit urna. Eget et ut quam malesuada congue.",
+    },
+    options: [
+        { title: "1ere", description: "Plateforme entièrement dédiée à l’actualité du football africain." },
+        { title: "+10", description: "Millions de magazine sur le foot africain diffusées." },
+        { title: "34", description: "Pays du monde pays comme point de diffusion." },
+        { title: "56", description: "Jurys issus de tout le contient pour choisir l’étoile d’or africain." }
+    ],
+    ourGoals: {
+
+    },
+    ourTeams: {
+        title: "Notre equipe de direction",
+        article: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Dignissim blandit arcu nisl erat sit urna. Eget et ut quam malesuada congue.",
+        asides: [
+            { image: person1, title: "NOM & PRENOM", job: "FONCTION (Ex: CEO)" },
+            { image: person2, title: "NOM & PRENOM", job: "FONCTION (Ex: CEO)" },
+            { image: person3, title: "NOM & PRENOM", job: "FONCTION (Ex: CEO)" },
+            { image: person4, title: "NOM & PRENOM", job: "FONCTION (Ex: CEO)" }
+        ]
+    },
+    ourFriends: {
+        title: "Notre equipe de direction",
+        article: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Dignissim blandit arcu nisl erat sit urna. Eget et ut quam malesuada congue.",
+        asides: [
+            { image: person1, title: "NOM & PRENOM", job: "FONCTION (Ex: CEO)" },
+            { image: person2, title: "NOM & PRENOM", job: "FONCTION (Ex: CEO)" },
+            { image: person3, title: "NOM & PRENOM", job: "FONCTION (Ex: CEO)" },
+            { image: person4, title: "NOM & PRENOM", job: "FONCTION (Ex: CEO)" }
+        ]
+    }
+}
+
+export default data
