+++ hemle_groupwork/src/components/About/style.css
@@ -1,126 +1,94 @@
-
-body { display:flex; justify-content:center; background:black; }
-
-#root { margin:0em; background:#1c1c1a; font-family:roboto; max-width:90em; overflow-x:hidden; }
-
-#vid { height:50%; margin:0em 2.5em; background-color:green; }
-
-
-.buttonchild { color:white; }
-
-#world { padding:2.5px; border-radius:5px; background:rgba(172,172,170,0.2); }
-
-
-article { position:absolute; width:710px; }
-
-article h1 { text-transform:uppercase; font-weight:100; font-size:40px; margin-bottom:20px; color:white; }
-
-article h2 { color:#FFFF27; font-style:oblique; font-size:28px; }
-
-article p { font-size:18.6px; line-height:30px; color:white; }
-
-
-
-section { position:relative; display:flex; overflow:hidden; }
-
-section img { transition:0.5s; position:absolute; }
-
-
-
-#sec1 { display:block; height:70em; }
-
-#sec1 p { line-height:22px; }
-
-#sec1 #art_africa { top:4.75em; left:5.5em; width:40em; }
-
-#sec1 #art_africa h1 { font-weight:bold; font-size:38px; }
-
-#sec1 #africa { top:1.5em; right:2em; }
-
-#sec1 #man { top:20em; left:-12em; }
-
-#sec1 #art_man { top:32em; right:5em; width:30em; }
-
-#sec1 #art_man h1 { font-weight:bold; font-size:32px; }
-
-#sec1 #art_man p { font-size:16px; }
-
-#sec1 #africa:hover, #sec1 #man:hover { scale:1.1; }
-
-#sec1 #art_man span { color:#008b3b; text-decoration:underline; }
-
-
-
-#sec2 { height:19em; background:url("../assets/home/section2.png") center; justify-content:center;
-    align-items:center; }
-
-
-
-    
-#sec3 { height:60em; overflow:hidden; margin-bottom:4em; }
-
-#lamp, #cam { position:absolute; }
-
-#lamp { top:0em; left:13em; transform:rotate(-31.84deg); height:740px; }
-
-#cam { top:12em; left:40em; transform:rotate(17.08deg); height:380px; }
-
-#sec3 img:hover { scale:1.1; }
-
-
-#objectifs { position:absolute; bottom:0; width:100%; height:20em; display:flex; justify-content:
-    space-between; align-items:center; flex-wrap:wrap; overflow:hidden; background:green;
-    margin-left:6em; }
-
-
-
-
-#article { width:40em; overflow:hidden; background:red; }
-
-#article h1 { text-transform:none; font-weight:normal; font-size:35px; font-weight:bold; }
-
-#article span { text-decoration:underline; }
-
-
-#asides { display:flex; background:blue; }
-
-#asides .asid { width:250px; height:280px; border-radius:15px; transform:skew(-37deg);
-    margin:0em 0.5em; overflow:hidden; }
-
-#asides .asid img { transform:skew(37deg); margin-left:-7em; }
-
-#asides .asid img:hover { scale:1.2; }
-
-
-
-
-
-#sec4 { flex-direction:column; align-items:center; }
-
-
-#sec4 #sec_info { width:35.25em; height:10em; display:flex; flex-direction:column; justify-content:space-around;
-    color:white; text-align:center; margin:5em 0; }
-
-#sec_info h1 { text-transform:uppercase; font-size:34px; letter-spacing:-1.5px; }
-
-#sec_info .img { display:flex; justify-content:center; }
-
-#sec_info p { line-height:30px; }
-
-
-#sec4 #allasid { display:flex; width:100%; height:400px; overflow-x:auto; }
-
-#allasid .asid { min-width:190px; height:400px; border-radius:8px; transform:skew(-23deg);
-    margin:0em 2em; overflow:hidden; background-color:green; }
-
-#allasid .asid img { transform:skew(23deg); margin-left:-85px; background-color:red; }
-
-#allasid:hover .asid img { scale:1.2; }
-
-
-
-
-footer { background:#d9d9d9; height:10.5em; overflow-x:auto; margin-bottom:5em;
-    display:flex; align-items:center; padding-bottom:2em; }
-
-footer img { margin:0.5em 1em; }
+body {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  
+  background: black;
+}
+
+#root {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  background: #292A2A;
+  width: 100%;
+  max-width: 90em;
+  padding-bottom: 4em;
+  color: black;
+}
+
+section {
+  display: flex;
+  background: #EAECE9;
+  position: relative;
+  padding: 3em 5% 0 5%;
+  width: 90%;
+  padding-top: 4em;
+  margin: 0;
+}
+
+section .title h2 {
+  text-transform: uppercase;
+  font-weight: bold;
+  font-size: 34px;
+  line-height: 67px;
+  color: black;
+}
+
+section .content, section .images {
+  display: flex;
+}
+
+section .article {
+  margin-top: 1em;
+  font-size: 18px;
+}
+
+section .article p {
+  margin: 1em 0;
+  line-height: 1.5em;
+}
+
+section .article a {
+  color: #009741;
+  text-decoration: underline;
+}
+
+@media all and (max-width: 50em) {
+  .content {
+    justify-content: center;
+  }
+}
+
+@media all and (max-width: 600px) {
+  section {
+    width: 96%;
+    padding-top: 1em;
+  }
+  section .title h2 {
+    font-size: 1.3em;
+    line-height: 37px;
+  }
+  section .title img {
+    width: 60px;
+  }
+  section .article {
+    font-family: "Roboto";
+    font-weight: 400;
+    font-size: 15.75px;
+    line-height: 25px;
+    margin:1em 0;
+  }
+}
+
+@media all and (max-width: 360px) {
+  section .title h2 {
+    font-size: 1.25em;
+    margin-bottom: -0.5em;
+  }
+  #root {
+    padding-bottom: 1em;
+  }
+}
