+++ hemle_groupwork/src/components/Cartes/Section/index.js
@@ -1,18 +1,15 @@
 import Card from "../Card"
 import "./style.css"
 
-const Section = ({section}) => {
-    let players = section.asides.map( tmp =>
-        <Card player={tmp}/>
-    )
-    return (
-        <section>
-            <h2> {section.title} </h2>
-            <div className="asides">
-                {players}
-            </div>
-        </section>
-    )
-}
+const Section = ({section}) => (
+    <section className="players_position">
+        <h2> {section.position} </h2>
+        <div className="asides">
+            { section.players.map( (tmp, i) =>
+                <Card player={tmp} key={i} />
+            ) }
+        </div>
+    </section>
+)
 
 export default Section
