+++ hemle_groupwork/src/components/Cartes/data.jsx
@@ -1,67 +1,66 @@
+import img from "../../assets/tools/player.png"
 
-import img from "./assets/person2.png"
-
-const theCards = {
-    gardiens: [
-        {
-            img: img,
-            number: 1,
-            shortName: "Pierre Onana",
-            job: "Defenseur central",
-            fullName: "Arnaud Boris Mbala Bengono"
-        },
-        {
-            img: img,
-            number: 2,
-            shortName: "Pierre Onana",
-            job: "Defenseur central",
-            fullName: "Arnaud Boris Mbala Bengono"
-        },
-        {
-            img: img,
-            number: 999,
-            shortName: "Pierre Onana",
-            job: "Defenseur central",
-            fullName: "Arnaud Boris Mbala Bengono"
-        },
-        {
-            img: img,
-            number: 999,
-            shortName: "Pierre Onana",
-            job: "Defenseur central",
-            fullName: "Arnaud Boris Mbala Bengono"
-        },
-        {
-            img: img,
-            number: 999,
-            shortName: "Pierre Onana",
-            job: "Defenseur central",
-            fullName: "Arnaud Boris Mbala Bengono"
-        },
-        {
-            img: img,
-            number: 999,
-            shortName: "Pierre Onana",
-            job: "Defenseur central",
-            fullName: "Arnaud Boris Mbala Bengono"
-        },
-    ],
-    attaquant: [
-        {
-            img: img,
-            number: 999,
-            shortName: "Pierre Onana",
-            job: "Defenseur central",
-            fullName: "Arnaud Boris Mbala Bengono"
-        },
-        {
-            img: img,
-            number: 999,
-            shortName: "Boris Mbala",
-            job: "Boris Mbala",
-            fullName: "Arnaud Boris Mbala Bengono",
-        },
-    ]
-}
+const theCards = [
+    {
+        position: "gardiens",
+        hg: "44",
+        players: [
+            {
+                img: img,
+                number: 1,
+                shortName: "Pierre Onana",
+                job: "Defenseur central",
+                fullName: "Arnaud Boris Mbala Bengono"
+            }, {
+                img: img,
+                number: 2,
+                shortName: "Pierre Onana",
+                job: "Defenseur central",
+                fullName: "Arnaud Boris Mbala Bengono"
+            }, {
+                img: img,
+                number: 999,
+                shortName: "Pierre Onana",
+                job: "Defenseur central",
+                fullName: "Arnaud Boris Mbala Bengono"
+            }, {
+                img: img,
+                number: 999,
+                shortName: "Pierre Onana",
+                job: "Defenseur central",
+                fullName: "Arnaud Boris Mbala Bengono"
+            }, {
+                img: img,
+                number: 999,
+                shortName: "Pierre Onana",
+                job: "Defenseur central",
+                fullName: "Arnaud Boris Mbala Bengono"
+            }, {
+                img: img,
+                number: 999,
+                shortName: "Pierre Onana",
+                job: "Defenseur central",
+                fullName: "Arnaud Boris Mbala Bengono"
+            }
+        ]
+    }, {
+        position: "attaquant",
+        players: [
+            {
+                img: img,
+                number: 999,
+                shortName: "Pierre Onana",
+                job: "Defenseur central",
+                fullName: "Arnaud Boris Mbala Bengono"
+            }, {
+                img: img,
+                number: 999,
+                shortName: "Boris Mbala",
+                job: "Boris Mbala",
+                fullName: "Arnaud Boris Mbala Bengono",
+            }
+        ]
+    }
+]
 
 export default theCards
