+++ hemle_groupwork/src/components/Cartes/style.css
@@ -1,9 +1,3 @@
-@font-face {
-    font-family: 'roboto';
-    src: url('./fonts/Roboto-Regular.ttf') format("TrueType");
-}
-  
 body {
     background: #3C3C3C;
-    font-family: roboto;
 }
