+++ hemle_groupwork/src/components/Infos/style.css
@@ -1,0 +1,193 @@
+body {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    
+    background: #272724;
+    color: white;
+}
+
+
+#biographie {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    max-width: 50em;
+    padding: 0 1em;
+}
+
+
+
+#biographie #infoSection {
+    display: flex;
+    justify-content: center;
+    align-items: flex-end;
+    align-self: center;
+    flex-wrap: wrap;
+    
+    width: 100%;
+}
+
+
+
+#biographie #infoSection .image {
+    display: flex;
+    justify-content: center;
+
+    width: 319px;
+    height: 311px;
+    overflow: hidden;
+    position: relative;
+    box-sizing: content-box;
+    margin-right: 2em;
+}
+
+#biographie #infoSection .image img {
+    /* width: 100%; */
+    margin-top: 3em;
+    transition: 1s;
+}
+
+#biographie #infoSection .image .controls {
+    position: absolute;
+    inset: 0;
+    background:
+        linear-gradient(#0000 0%, rgba(0, 0, 0, 0.76) 76.56%, #1F1414 100%);
+}
+
+#biographie #infoSection .image .number {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    background: url("./assets/circle.png");
+    background-size: 100% 100%;
+    width: 55px; height: 55px;
+    border-radius: 50%;
+    margin: 28px 20px;
+}
+
+
+
+
+
+#biographie #infoSection .informations {
+    flex: 1;
+    max-width: 372px;
+    background: #292A2A;
+}
+
+#biographie .informations .infosDiv {
+    display: flex;
+    flex-direction: column;
+
+    width: 319px;
+    background: #292A2A;
+    padding: 0 1em;
+    color: white;
+}
+
+#biographie .informations .title {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-wrap: wrap;
+
+    border-bottom: 1px solid white;
+    margin-bottom: 1em;
+}
+
+#biographie .informations .title img {
+    width: 3em;
+    height: 3em;
+    border-radius: 50%;
+    margin-right: 10px;
+}
+
+#biographie .informations .title h2 {
+    font-size: 18px;
+    padding: 25px 0;
+    text-align: center;
+}
+
+
+#biographie .informations .infosDiv .label {
+    font-size: 14px;
+    font-weight: 600;
+    line-height: 45px;
+    text-transform: capitalize;
+}
+
+#biographie .informations .infosDiv .value {
+    font-size: 15px;
+}
+
+
+
+
+#biographie #infoSection:hover .image img {
+    scale: 1.15;
+}
+#biographie #infoSection:hover .informations h2 {
+    transition: 0.7s;
+    color: yellow;
+}
+
+
+
+
+
+
+#biographie #playerDescription {
+    padding: 1.5em 0;
+    width: 100%;
+}
+
+#biographie #playerDescription h2 {
+    display: flex;
+    align-items: center;
+
+    text-transform: uppercase;
+    font-size: 15px;
+}
+
+#biographie #playerDescription h2::after {
+    content: '';
+    display: block;
+    width: 12px;
+    height: 2px;
+    margin-left: 5px;
+
+    background: linear-gradient(90deg, #DF251A 6.4%, #E02D1A 12.32%, #E44419 21.71%,
+    #E96A17 33.39%, #F19D16 46.8%, #F7C914 56.97%, #ECC716 60.02%, #CDC01B 65.45%,
+    #9CB624 72.62%, #58A730 81.26%, #039540 91.05%, #009440 91.32%);
+}
+
+
+#biographie #playerDescription p {
+    font-size: 15px;
+    line-height: 20px;
+    box-sizing: border-box;
+    padding: 20px 1em;
+}
+
+
+
+#biographie @media all and (max-width: 890px) {
+    #biographie #infoSection .image {
+        width: 100%;
+        margin-right: 0;
+    }
+    #biographie #infoSection .image img {
+        object-fit: contain;
+    }
+    #biographie #infoSection .informations {
+        max-width: 100%;
+    }
+    #biographie #infoSection .image .number {
+        margin: 0;
+        width: 8px;
+        height: 8px;
+    }
+}
