+++ hemle_groupwork/src/style.css
@@ -4,97 +4,6 @@
 }
 
 body {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  
-  background: black;
   min-height: 100vh;
   font-family: "Roboto";
-}
-
-#root {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  background: #292A2A;
-  width: 100%;
-  max-width: 90em;
-  padding-bottom: 4em;
-}
-
-section {
-  display: flex;
-  background: #EAECE9;
-  position: relative;
-  padding: 3em 5% 0 5%;
-  width: 90%;
-  padding-top: 4em;
-  margin: 0;
-}
-
-section .title h2 {
-  text-transform: uppercase;
-  font-weight: bold;
-  font-size: 34px;
-  line-height: 67px;
-  color: black;
-}
-
-section .content, section .images {
-  display: flex;
-}
-
-section .article {
-  margin-top: 1em;
-  font-size: 18px;
-}
-
-section .article p {
-  margin: 1em 0;
-  line-height: 1.5em;
-}
-
-section .article a {
-  color: #009741;
-  text-decoration: underline;
-}
-
-@media all and (max-width: 50em) {
-  .content {
-    justify-content: center;
-  }
-}
-
-@media all and (max-width: 600px) {
-  section {
-    width: 96%;
-    padding-top: 1em;
-  }
-  section .title h2 {
-    font-size: 1.3em;
-    line-height: 37px;
-  }
-  section .title img {
-    width: 60px;
-  }
-  section .article {
-    font-family: "Roboto";
-    font-weight: 400;
-    font-size: 15.75px;
-    line-height: 25px;
-    margin:1em 0;
-  }
-}
-
-@media all and (max-width: 360px) {
-  section .title h2 {
-    font-size: 1.25em;
-    margin-bottom: -0.5em;
-  }
-  #root {
-    padding-bottom: 1em;
-  }
 }
